+++ src/main/java/transaction/TransactionList.java
@@ -3,7 +3,6 @@ package transaction;
 import car.CarList;
 import exceptions.CarException;
 import parser.CarParser;
-
 import java.util.ArrayList;
 
 public class TransactionList {
@@ -24,7 +23,11 @@ public class TransactionList {
         System.out.println("Transaction added: ");
         System.out.println(transaction);
     }
-          
+
+    public static void addTxWithoutPrintingInfo(Transaction transaction) {
+        transactionList.add(transaction);
+    }
+
     public static void printAllTransactions() {
         int index = 1;
 
@@ -119,4 +122,18 @@ public class TransactionList {
         }
         System.out.println("Transaction not found");
     }
+
+    public static String transactionListToFileString(){
+        StringBuilder transactionData = new StringBuilder();
+        for (Transaction transaction : transactionList) {
+            transactionData.append(transaction.toFileString());
+            transactionData.append("\n");
+        }
+        return transactionData.toString();
+    }
+
+    public static ArrayList<Transaction> getTransactionList(){
+        return transactionList;
+    }
+
 }
