+++ src/test/java/file/TransactionFileTest.java
@@ -2,8 +2,8 @@ package file;
 
 
 import org.junit.jupiter.api.Test;
-import transcation.Transaction;
-import transcation.TransactionList;
+import transaction.Transaction;
+import transaction.TransactionList;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class TransactionFileTest {
@@ -25,17 +25,17 @@ class TransactionFileTest {
         parameters = new String[]{"12313" ,"ohn" ,"me" ,"2024-10-29"};
         transactionFile.addTransactionWithParameters(parameters);
         Transaction transaction1 = TransactionList.getTransactionList().get(0);
-        assertEquals(transaction1.getBorrowerName(), "John");
+        assertEquals(transaction1.getCustomer(), "John");
         assertEquals(transaction1.getCarLicensePlate() , "SBS123B");
         assertEquals(transaction1.getDuration(), "3");
         assertEquals(transaction1.getStartDate() , "2024-10-17");
         Transaction transaction2 = TransactionList.getTransactionList().get(1);
-        assertEquals(transaction2.getBorrowerName(), "Jon");
+        assertEquals(transaction2.getCustomer(), "Jon");
         assertEquals(transaction2.getCarLicensePlate() , "SB");
         assertEquals(transaction2.getDuration(), "0");
         assertEquals(transaction2.getStartDate() , "2024-11-17");
         Transaction transaction3 = TransactionList.getTransactionList().get(2);
-        assertEquals(transaction3.getBorrowerName(), "ohn");
+        assertEquals(transaction3.getCustomer(), "ohn");
         assertEquals(transaction3.getCarLicensePlate() , "12313");
         assertEquals(transaction3.getDuration(), "me");
         assertEquals(transaction3.getStartDate() , "2024-10-29");
